+++ CHANGELOG.md
@@ -3,6 +3,27 @@
 All notable changes to this project will be documented in this file. See
 [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+### [2.12.5](https://github.com/seamapi/python/compare/v2.12.4...v2.12.5) (2023-06-06)
+
+
+### Bug Fixes
+
+* Update events methods params ([#93](https://github.com/seamapi/python/issues/93)) ([3743539](https://github.com/seamapi/python/commit/37435391dd198886087662940c39b6174487426b))
+
+### [2.12.4](https://github.com/seamapi/python/compare/v2.12.3...v2.12.4) (2023-06-06)
+
+
+### Bug Fixes
+
+* Update Provider Category Property Name ([#97](https://github.com/seamapi/python/issues/97)) ([9c0b3d2](https://github.com/seamapi/python/commit/9c0b3d27142bf3067465b237b8be4c9ae1543d76))
+
+### [2.12.3](https://github.com/seamapi/python/compare/v2.12.2...v2.12.3) (2023-06-05)
+
+
+### Bug Fixes
+
+* Add Category to Connect Webview Create ([#96](https://github.com/seamapi/python/issues/96)) ([f82be14](https://github.com/seamapi/python/commit/f82be144ad96c9e439a7e222b15eafe023cdb1ad))
+
 ### [2.12.2](https://github.com/seamapi/python/compare/v2.12.1...v2.12.2) (2023-05-25)